improving models

+++ server/db/user.ts
@@ -3,14 +3,12 @@ import { prisma } from '.'
 
 export function createUser(data: { username: string, email: string, password: string }) {
   const hashPassword = bcrypt.hashSync(data.password, 10)
-  const verificationCode = crypto.randomUUID()
 
   return prisma.user.create({
     data: {
       username: data.username,
       email: data.email,
       password: hashPassword,
-      verificationCode,
     },
   }) // throws P2002 if email already exists
 }
@@ -25,51 +23,41 @@ function findUserNotVerifiedById(id: string) {
 }
 
 export function deleteUserById(id: string) {
-  return new Promise((resolve, reject) => {
-    prisma.refreshToken
-      .deleteMany({
-        where: {
-          uid: id,
-        },
-      })
-      .then(async () => {
-        await prisma.user
-          .delete({
-            where: { id },
+  return new Promise<true>((resolve, reject) => {
+    // Delete user sessions
+    prisma.session.deleteMany({ where: { userId: id } })
+      .then(() => {
+        // Delete user verification code
+        prisma.verificationCode.delete({ where: { userId: id } })
+          .then(() => {
+            // Delete user
+            prisma.user.delete({ where: { id } })
+              .then(() => resolve(true))
+              .catch(error => reject(error))
           })
-          .catch((error) => {
-            reject(error)
-          })
+          .catch(error => reject(error))
       })
-      .catch((error) => {
-        reject(error)
-      })
-
-    resolve(true)
+      .catch(error => reject(error))
   })
 }
 
 export function deleteUserByIdIfNotVerified(id: string, timeout: number) {
-  return new Promise((resolve, reject) => {
+  return new Promise<true>((resolve, reject) => {
     setTimeout(async () => {
+      // Find the user by id if its not verified
       await findUserNotVerifiedById(id)
-        .then(async () => {
-          await deleteUserById(id)
-            .then(() => {
-              resolve(true)
-            })
-            .catch((error) => {
-              reject(error)
-            })
+        .then(() => {
+          // Delete the user
+          deleteUserById(id)
+            .then(() => resolve(true))
+            .catch(error => reject(error))
         })
         .catch((error) => {
-          if (error.message === 'No User found')
-            resolve(true)
           reject(error)
         })
     }, timeout)
   })
-};
+}
 
 export function findUserById(id: string) {
   return prisma.user.findUniqueOrThrow({ where: { id } }) // throws P2025 if not found
@@ -80,7 +68,6 @@ export function verifyUser(id: string) {
     where: { id },
     data: {
       verified: true,
-      verificationCode: null,
     },
   })
 }
